feat: Add test form
Refs: #RW-1244
