+++ html/modules/custom/reliefweb_import/reliefweb_import.services.yml
@@ -10,3 +10,7 @@ services:
   reliefweb_moderation.reliefweb_import.moderation:
     class: Drupal\reliefweb_import\Service\ReliefWebImporterModeration
     arguments: ['@current_user', '@database', '@date.formatter', '@entity_field.manager', '@entity_type.manager', '@pager.manager', '@pager.parameters', '@request_stack', '@string_translation', '@plugin.manager.reliefweb_import.reliefweb_importer']
+
+  reliefweb_import.inoreader_service:
+    class: 'Drupal\reliefweb_import\Service\InoreaderService'
+    arguments: ['@http_client', '@state']
